I created a new class called ResultsFragment which communicates with the "MainScreenActivity". It is enabled when the toggle switch is checked off. Right now the login screen isn't working so I am stopping this work for now
